+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 1.5.18 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.
+## New Modern Content:
+- Studdied Attack
+- Summer Winds
 ## Bug Fixes:
 - Emboldening Bond
 - Protective Bond